refactor: Rename folder from QuestionnaireLabel to Questionnaire
