+++ src/components/common/Questionnaire/CollapsedDrawer/CollapsedDrawer.module.css
