+++ src/components/common/Questionnaire/ExpandedModal/AppreciationContent/flower.png
