+++ src/components/common/Questionnaire/ExpandedModal/laborImage.png
